feat: 서준호 - 프로젝트이력 cud 메소드정의. - 프로젝트 선택하는 팝업창 구현해야함

+++ Winplus_SIS/src/main/java/com/samyang/winplus/common/system/authority/service/SystemAuthorityServiceImpl.java
@@ -434,7 +434,20 @@ public class SystemAuthorityServiceImpl implements SystemAuthorityService {
 				} else if ("U".equals(crud)){
 					resultInt += systemAuthorityDao.updateEmp(paramMap);
 				} else if ("D".equals(crud)){
-					resultInt += systemAuthorityDao.deleteEmp(paramMap);
+					
+					
+					//하위 메뉴 존재여부 체크
+					int chkChildPjtInt = systemAuthorityDao.chkChildPjt(paramMap);
+					
+					
+					//하위메뉴 존재 시 메뉴 삭제 안됨
+					if(chkChildPjtInt != 0){
+						chkChildPjtInt = 9999;
+						return chkChildPjtInt;
+					}else{
+						resultInt += systemAuthorityDao.deleteEmp(paramMap);
+					}
+				
 				}
 			}
 		}
@@ -477,6 +490,7 @@ public class SystemAuthorityServiceImpl implements SystemAuthorityService {
 				} else if ("U".equals(crud)){
 					resultInt += systemAuthorityDao.updatePjt(paramMap);
 				} else if ("D".equals(crud)){
+					
 					resultInt += systemAuthorityDao.deletePjt(paramMap);
 				}
 			}
